install: Update mega instructions
This little script checks whether the current upstream lsb release (e.g.
the base ubuntu version for a linux mint distribution) has an existing mega
package url.
Note that, wget will fail otherwise and stop the execution of the script
due to the 'set -e' in the python script.
It is totally fine if this updating fails, since the user can relatively
easily fix this manually.