sysd+X+pam: Solve age long mystery of environments
For a very long time I was puzzled by how the "boot" process for the
user works and how the environment variables are affected.
I've had numerous attempts at changing the environment for ALL processes
from a given user but failed again and again. Eventually, I got the
systemd "root" process correctly configured, but the children - namely
sxhkd - did not inherit this environment... In the end, the solution was
so simple: The After= directive is mandatory.