sysd: Add firefox profile backup just like chrome
Ideally, I would just version control the user.js file and other things
like passwords and... Well, I don't quite know how much I care about all
the other configs, it would be nice not to authenticate to every site
when install the config on a new system.. I need to think about it
again. Until then, this is the easiest solution for me to implement.