firefox: Copy over auto-profile and picker from chrome
firefox-auto-profile tries to find a firefox instance running on the
current workspace. If found, it will launch subsequent links in said
instance (more specifically, in an instance with the same profile).
This requires the previous commit as well as changes in mimeapps.list,
which will follow soon.
If no instance could be found, or no profile was specified for that
instance, the main profile will be used.

firefox-profile-picker is needed despite the firefox -ProfileManager,
because this does not change the default profile which is supposed to be
main.