touch: Add skip-youtube-ads script
This little script (interpreted by xdotool) dismisses all youtube ads
for a Full-HD screen while making sure the video (in full screen mode) keeps playing.

The order of the commands and the amount of clicks is crucial.
The assumption is that this shortcut will only be invoked if there is
actually one of the ads showing.

If Banner 1 or Banner 2 is displayed, the screen will be clicked 4 times
(apart from dismissing the ad itself). Since this is an even number, the
video will still be playing after the script if it was before, and
vice-versa.
If there is a video ad showing, the clicks for Banner 1 and 2 don't have
an impact on the actual youtube video playing and Random click and
Survey will keep the play-state.
The survey case is not that complicated.

Please note that it is necessary to click on X=1919 so that we don't
accidentally hit the timeline and skip forward to the end of the video.