vim: Add autochmod: autocmds to chmod before write
This very primitive plugin is supposed to automate the changing
of permissions of files that you edited, but don't have permissions to
write to. In neovim, using sudo like this :w !sudo tee % doesn't work.

I wouldn't recommend this approach to anybody else, since editing files
you're not allowed to is an anti-pattern anyways and I also did not pay
much attention to security. This will probably be removed once there is
a cleaner way to do this in neovim.