shell: Don't use cd -P for aliases .. ... ....
In my version of zsh, cd -P -L does not work as expected. I thought it
shouldn't follow symlinks like cd -L does, essentially overwriting -P,
but it doesn't.. This works in bash however.
This means that you can't go "back" one directory if you followed a
symlink, not even with .. -L. The proper way would be to find out if
this is an ongoing bug and file a report, but I just use normal cd with
no arguments for my aliases .. ... .... for now.