ws: Add temporary fix for non-responsive firefox
When firefox is launched, it somehow assumes fills the whole screen.
This leaves half of the page inaccessible, if I open 2 windows
horizontally next to each other. Since I could not find another
solution (I didn't look too much), I added this simple hotfix.

It waits a little, then tries to focus the window which corresponds to
the firefox pid. After that, super+f is pressed twice, which puts
firefox in fullscreen and back, making forcing the page to be
rerendered.