vim: Mv <leader>l to <leader>o and don't submit line
This makes it so I can easily open another file and don't need two
shortcuts for these two use cases. I tried to write a function that would
leave you with the Ex command line open if the file under cursor does
not exist, but apparently it is not possible to start a 'dangling'
command.