vim: Add mappings for changing directories
This is useful when the file you're editing has a reference to some
other file relative to it's directory (for example in a README.md). If
you <leader>c before, you can use gf and all the other file finding
mechanisms of vim. Also handy if you started vim from the wrong dir
(like with my ^R of zle).