repo: Explicitly (un-)exclude files
This lets me see, if I have any untracked files that should be tracked
in specific directories. I was about to develop a tool that checks said
directories for untracked files, because I quite liked the
elegant status.showUntrackedFiles config option. This is not as clean,
but it does what I want.