vim+cd: Use eunuch to replace my AutoChmod
In the referenced Post, I realised that handling password saving inside
Vim is probably not a good idea. I'm not a security expert and Vim is
(probably?) not a hardened secure application. I really wanted to be
prompted through the terminal, but I guess a little graphical program is
fine..

See https://github.com/tpope/vim-eunuch/issues/56#issuecomment-439670169