ws: Explicitly name target tmux session everytime
I had to change this because of my new tmux launcher. When not inside a
tmux session, tmux will use the last referenced session as the default target
session. When inside a session, it is used everytime no matter which
session was referenced last by a command.
The new-window -t [shell-command] trick allows all commands inside the
script to run in the target session, without the need to explicitly
state in in every of these commands.