env: Use .profile to set envs
I started using .pam_environment for DCONF_PROFILE, which was used
before .profile was invoked. That's why I kept all the other envs in
there too. Now that I don't use DCONF_PROFILE anymore, there is no need to set
the other ones that early in the boot process and .profile is a more
appropriate place for them.