install: Use a shell script instead of py+yaml
This makes the install file(s) even smaller and more simple. I held on
the python script because of the logging feature, but in the end, you
only really need logging when there is something that's gone wrong. Thus
I added a small manual stacktrace so I can debug the error message,
should there be one.