zsh: Fix func
The unfunction command takes in the function name but does not accept a
full pathname to the function file, which is what was passed in before.

Another issue was the used subshell. I think I used it because of the
`set -eu`, but it unfunctioned the function in the wrong context.